Split models.py into a package along resource boundaries
Mirror the views/ layout: auth keys, environments, agents, and sessions
each get their own module, with models/__init__.py re-exporting so all
existing imports continue to work unchanged. Session FKs now use the
"fairy.<Model>" explicit app label so they stay decoupled from package
import order.
